Ex1/requirements.txt
flask==3.1.1
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
openai==1.78.0
pyodbc==5.2.0
sqlalchemy==2.0.40
pypdf==5.4.0
faiss-cpu==1.10.0
python-dotenv==1.1.0
